Dev Mode now defaults to dev server behavior
